- db 모델 데이터형 변경 - history dao추가 - history firestore에서 가져오도록 수정 - 필요없는 클래스 삭제

+++ app/src/main/java/com/nexters/yetda/android/price/PriceActivity.kt
@@ -32,8 +32,8 @@ class PriceActivity : BaseActivity<ActivityPriceBinding, PriceViewModel>() {
     override fun initDataBinding() {
         viewModel.startNextActivityEvent.observe(this, Observer {
             //save realm object to History
-            history.startPrice = leftValue
-            history.endPrice = rightValue
+            history.startPrice = leftValue.toLong()
+            history.endPrice = rightValue.toLong()
             startActivity(Intent(applicationContext, QuestionActivity::class.java))
         })
         viewModel.backBeforeActivityEvent.observe(this, Observer {
